Handle additional format for size metadata
There are other ways ffmpeg outputs the size than just \d+x\d+ followed
by a comma or newline.  There is a case where ffmpeg will output
additional metadata before the comma/newline in the form of
"600x338 [SAR 845:843 DAR 500:281]".  This does not work with the parser
as is, and will result in no size metadata being output which can be
problematic for anything relying on it being there.